Crate shared_canvas

|_________| is human nature. You, and you alone, are responsible for your own actions. With great power comes great responsibility. You may choose to use your power to work with others and make something beautiful, or you may choose to use your power for selfish purposes. Choose wisely.

This shared canvas uses the universe-atom model implemented in sharing_is_caring. See that crate for more details.

Modules

canvas
Contains implementation for a shared canvas. Interfaces with the sharing_is_caring model, where the canvas is the universe, and each pixel is an atom.

responder
Contains implementation for responding to client requests.