Module shared_canvas::canvas

Contains implementation for a shared canvas. Interfaces with the sharing_is_caring model, where the canvas is the universe, and each pixel is an atom.

Structs

Canvas
Canvas struct implements the server side's implementation of the canvas. It keeps track of the width, height, pixels, and the pixel_size to be used to draw the canvas on the client-side. This is the universe in the sharing_is_caring model.

Pixel
The Pixel struct represents a single pixel (square) on the canvas.