Struct shared_canvas::canvas::Canvas

pub struct Canvas { /* fields omitted */ }

Canvas struct implements the server side's implementation of the canvas. It keeps track of the width, height, pixels, and the pixel_size to be used to draw the canvas on the client-side. This is the universe in the sharing_is_caring model.

Methods

impl Canvas

pub fn new(width: usize, height: usize, pixel_size: usize) -> Self

Default constructor.